edit error NewsHomeList
