+++ src/components/home/NewsHomeList.jsx
@@ -1,28 +1,50 @@
-import React, { useState, useEffect } from 'react';
-import { Box, SimpleGrid } from '@chakra-ui/react';
+import React, { useState, useEffect } from 'react'
+import { Box, SimpleGrid } from '@chakra-ui/react'
 import { getAllEntries } from '../../services/entriesService'
-import NewsHomeCard from '../news/NewsHomeCard';
+import NewsHomeCard from '../news/NewsHomeCard'
+import Spinner from '../Spinner'
 
 const ListNews = () => {
-  const [newss, setNews] = useState([])
+  const [allNews, setNews] = useState([])
+  const [error, setError] = useState(null)
 
-  useEffect(() => {
+  const loadData = async () => {
     getAllEntries().then((entries) => {
       setNews(entries.data.body)
-      // eslint-disable-next-line no-console
-    }).catch((error) => console.log(error))
+      setError(null)
+    }).catch((err) =>
+      setError(err))
+  }
+
+  // run on load
+  useEffect(() => {
+    loadData()
   }, [])
 
-  return (
-    <Box p={4}>
-      <SimpleGrid columns={{ base: 1, md: 4 }} spacing={6}>
+  const getListItems = () =>
+    (
+      <Box p={4}>
+        <SimpleGrid columns={{ base: 1, md: 4 }} spacing={6}>
 
-        {
-        newss.map((news) =>
-          <NewsHomeCard {...news} key={news.id} />).reverse().slice(0, 4)
-        }
-      </SimpleGrid>
-    </Box>
+          { !allNews ? <Spinner /> : allNews.map((news) =>
+            <NewsHomeCard {...news} key={news.id} />).reverse().slice(0, 4)}
+        </SimpleGrid>
+      </Box>
+    )
+
+  const getErrorView = () =>
+    (
+      <div>
+        Error al cargar las novedades!
+      </div>
+    )
+
+  return (
+    <div>
+      <ul>
+        { error ? getErrorView() : getListItems() }
+      </ul>
+    </div>
   )
 }
