+++ src/pages/Home.js
@@ -1,14 +1,13 @@
 import React from 'react'
 
-import WelcomeText from '../components/WelcomeText'
-import NewsHome from '../components/NewsHome'
-import Slider from '../components/Slider'
+import WelcomeText from '../components/home/WelcomeText'
+import NewsHome from '../components/home/NewsHome'
+// import Slider from '../components/Slider'
 
 export default function Home() {
   return (
     <>
-      <h1>Hola Home</h1>
-      <Slider />
+      {/* <Slider /> */}
       <WelcomeText />
       <NewsHome />
     </>
