--- src/pages/Home.jsx
@@ -1,15 +1,0 @@
-import React from 'react'
-// import Slider from './Slider'
-import WelcomeText from '../components/home/WelcomeText'
-import NewsHome from '../components/home/NewsHome'
-
-const Home = () =>
-  (
-    <>
-      {/* <Slider /> */}
-      <WelcomeText />
-      <NewsHome />
-    </>
-  )
-
-export default Home
